// src/modules/order/types/OrderDetail.ts
export type OrderDetail = {
  name: string;
  imageUrl: string;
  quantity: string;
  cancelled: string;
  measurementUnit: string;
  price: string;
  productDiscount: string;
  seller: string;
  currentStateDescription: string;
  deliveryDate: string;
  skuCode: string;
  orderDetail: OrderDetail[]
};